requirements.txt
tensorflow-cpu==2.11.0
Pillow==8.0.1
matplotlib==3.4.2
numpy==1.20.2
pytest==6.2.4
python-dotenv==0.19.2
